Added ratings filter and fixed price categories bug and search handling.
Now it filters by ratings w/ call to the backend. Additionally, fixed price categories bug so it now filters the prices correctly, and so that if all are deselected then all are chosen. Also, started set up process for search. There is now an event listener to listen for enter button and setKeywords to searchbar value.